Enhance MemberInfoExtensions and refactor tests
- Updated `MemberInfoExtensions` to handle indexer properties with a `NotSupportedException`.
