+++ test/sharp-meta.Tests/Data/SampleClass.cs
@@ -36,4 +36,6 @@ public class SampleClass
     /// </summary>
     [Obsolete("This method is obsolete")]
     public void ObsoleteMethod() { }
+
+    internal static object MethodWithoutDocComments() => throw new NotImplementedException();
 }
